feat(card): create high-level CardWithImage
